LayoutTests/pointer-lock/lock-already-locked-expected.txt

Test calling lock when already in a locked state.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.webkitPointerLockElement is null
     Locking targetdiv1.
PASS document.onwebkitpointerlockchange event received.
PASS document.webkitPointerLockElement is targetdiv1
     Locking targetdiv1, again.
PASS document.onwebkitpointerlockchange event received.
PASS document.webkitPointerLockElement is targetdiv1
     Locking targetdiv2.
PASS document.onwebkitpointerlockchange event received.
PASS document.webkitPointerLockElement is targetdiv2
PASS successfullyParsed is true

TEST COMPLETE
